pyproject.toml
[tool.poetry]
name = "flask-test"
version = "0.1.0"
description = ""
authors = ["Fellowshift <worldoftanks.kirya.2012@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-restful = "^0.3.10"
psycopg2 = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
